requirements.txt
anticaptchaofficial==1.0.64
asgiref==3.5.2
async-timeout==5.0.1
attrs==25.3.0
Automat==24.8.1
Brotli==1.0.9
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cprint==1.2.2
cryptography==38.0.4
cssselect==1.3.0
defusedxml==0.7.1
filelock==3.17.0
Flask==2.2.5
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.6.1
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kafka-python==2.0.6
kaitaistruct==0.10
ldap3==2.9.1
lxml==5.3.1
MarkupSafe==3.0.2
mitmproxy==9.0.1
mitmproxy_wireguard==0.1.23
msgpack==1.0.8
packaging==24.2
parsel==1.10.0
passlib==1.7.4
playwright==1.50.0
Protego==0.4.0
protobuf==4.25.6
publicsuffix2==2.20191221
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyee==12.1.1
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyperclip==1.8.2
queuelib==1.7.0
redis==5.2.1
requests==2.32.3
requests-file==2.1.0
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.12
Scrapy==2.12.0
scrapy-playwright==0.0.43
scrapy-redis==0.9.1
scrapy-rotating-proxies==0.6.2
scrapy-user-agents==0.1.1
service-identity==24.2.0
six==1.17.0
sortedcontainers==2.4.0
termcolor==2.5.0
tldextract==5.1.3
tls-client==1.0.1
tomli==2.2.1
tornado==6.4.2
Twisted==24.11.0
typing==3.7.4.3
typing_extensions==4.4.0
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.3.0
urwid==2.1.2
user-agents==2.2.0
w3lib==2.3.1
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.21.0
zope.interface==7.2
zstandard==0.19.0
fastapi
pydantic